exercises 4.15 - 4.23 done, check delete blog test, not working because users only can delete blogs associated to them

+++ backend/controllers/blogs.js
@@ -1,13 +1,18 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
+const User = require("../models/user")
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  
+  const user = request.user
 
   if (!body.title || !body.url) {
     return response.status(400).send({
@@ -19,21 +24,26 @@ blogsRouter.post('/', async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes === undefined ? 0 : body.likes
+    likes: body.likes === undefined ? 0 : body.likes,
+    user: user.id
   })
+
   savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user.id) {
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
-  } catch(exception) {
-    next(exception)
+  } else {
+    return response.status(401).json({ error: 'invalid user' })
   }
 })
-
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
